Split out TelescopeStateBase base class
It holds the bits that don't talk to the backend, and hence do not need
to be specialised for async versus sync communication.
